fix(inference): move network config classes out of model_registry into dedicated module
Create network_config.py with TLSConfig, ProxyConfig, TimeoutConfig,
LimitsConfig, and NetworkConfig as a standalone module. These are
semantically unrelated to model registry logic.

Update all source and test imports to pull network config classes from
the new module while keeping RemoteInferenceProviderConfig,
ProviderModelEntry, and ModelRegistryHelper in model_registry.py.

Keep backward-compatible re-exports in model_registry.py for existing
import paths. Regenerate provider documentation with updated field
descriptions.
