+++ src/ogx/providers/inline/interactions/impl.py
@@ -31,7 +31,7 @@ from ogx.core.datatypes import AccessRule
 from ogx.core.request_headers import get_authenticated_user
 from ogx.log import get_logger
 from ogx.providers.utils.inference.http_client import build_network_client_kwargs
-from ogx.providers.utils.inference.model_registry import NetworkConfig
+from ogx.providers.utils.inference.network_config import NetworkConfig
 from ogx.providers.utils.interactions.interactions_store import InteractionsStore
 from ogx_api import (
     Inference,
